+++ runtime/engine/src/test/java/io/aklivity/zilla/runtime/engine/internal/ReconfigureIT.java
@@ -22,6 +22,7 @@ import static org.junit.rules.RuleChain.outerRule;
 import java.io.FileOutputStream;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.CountDownLatch;
@@ -39,7 +40,7 @@ import io.aklivity.zilla.runtime.engine.test.EngineRule;
 import io.aklivity.zilla.runtime.engine.test.annotation.Configuration;
 
 
-public class ReconfigureModifyIT
+public class ReconfigureIT
 {
     private final K3poRule k3po = new K3poRule()
         .addScriptRoot("net", "io/aklivity/zilla/specs/engine/streams/network")
@@ -61,7 +62,7 @@ public class ReconfigureModifyIT
     @Rule
     public final TestRule chain = outerRule(engine).around(k3po).around(timeout);
 
-    private final String packageName = ReconfigureModifyIT.class.getPackageName();
+    private final String packageName = ReconfigureIT.class.getPackageName();
     private Path configDir = Paths.get("target/test-classes", packageName.replace(".", "/"));
 
     @Before
@@ -81,13 +82,12 @@ public class ReconfigureModifyIT
     })
     public void shouldReconfigureWhenModified() throws Exception
     {
-        System.out.println("Modify");
         k3po.start();
         k3po.awaitBarrier("CONNECTED");
 
         Path target = configDir.resolve("zilla.reconfigure.modify.json");
 
-        try (InputStream source = ReconfigureModifyIT.class.getResourceAsStream("zilla.reconfigure.after.json");
+        try (InputStream source = ReconfigureIT.class.getResourceAsStream("zilla.reconfigure.after.json");
              OutputStream outputStream = new FileOutputStream(target.toFile()))
         {
             source.transferTo(outputStream);
@@ -95,6 +95,45 @@ public class ReconfigureModifyIT
 
         EngineTest.TestEngineExt.registerLatch.await();
         k3po.notifyBarrier("CONFIG_CHANGED");
+
+        k3po.finish();
+    }
+
+    @Test
+    @Configuration("zilla.reconfigure.delete.json")
+    @Specification({
+        "${app}/client.sent.data.reconfigure.delete/server",
+        "${net}/client.sent.data.reconfigure.delete/client"
+    })
+    public void shouldReconfigureWhenDeleted() throws Exception
+    {
+        k3po.start();
+        k3po.awaitBarrier("CONNECTED");
+
+        configDir.resolve("zilla.reconfigure.delete.json").toFile().delete();
+        EngineTest.TestEngineExt.registerLatch.await();
+        k3po.notifyBarrier("CONFIG_DELETED");
+
+        k3po.finish();
+    }
+
+    @Test
+    @Configuration("zilla.reconfigure.missing.json")
+    @Specification({
+        "${app}/client.sent.data.reconfigure.create/server",
+        "${net}/client.sent.data.reconfigure.create/client"
+    })
+    public void shouldReconfigureWhenCreated() throws Exception
+    {
+        k3po.start();
+        try (InputStream source = ReconfigureIT.class.getResourceAsStream("zilla.reconfigure.original.json"))
+        {
+            Path target = configDir.resolve("zilla.reconfigure.missing.json");
+            Files.copy(source, target);
+        }
+
+        EngineTest.TestEngineExt.registerLatch.await();
+        k3po.notifyBarrier("CONFIG_CREATED");
 
         k3po.finish();
     }
